// src/app/models/enums/form-field-names/bom-configuration-form-field.ts
export enum BomConfigFormField {
  NAME = "name",
  PART_NUMBER_SHEET = "partNumberSheet",
  PART_NUMBER_COLUMN = "partNumberColumn",
  PART_NUMBER_ROW = "partNumberRow",
  REVISION_SHEET = "revisionSheet",
  REVISION_COLUMN = "revisionColumn",
  REVISION_ROW = "revisionRow",
  DESCRIPTION_SHEET = "descriptionSheet",
  DESCRIPTION_COLUMN = "descriptionColumn",
  DESCRIPTION_ROW = "descriptionRow",
  MATERIALS_SHEET = "materialsSheet",
  MATERIALS_START_ROW = "materialsStartRow",
  MFG_NAME_COLUMN = "mfgNameColumn",
  MFG_PN_COLUMN = "mfgPnColumn",
  MATERIALS_DESCRIPTION_COLUMN = "materialsDescriptionColumn",
  QUANTITY_COLUMN = "quantityColumn",
  UNIT_COLUMN = "unitColumn",
}